agent/requirements.txt
# ailive-realtime-agent dependencies
# 注意：LiveKit 套件全部鎖 ==1.5.1（江彬血淚教訓 #6：版本不一致會 ChunkedStream crash）

# LiveKit 核心
livekit>=1.0.0,<2.0.0
livekit-agents==1.5.1
livekit-plugins-elevenlabs==1.5.1
livekit-plugins-silero==1.5.1
livekit-plugins-anthropic==1.5.1
livekit-plugins-deepgram==1.5.1

# MiniMax 自訂 wrapper（不用官方 plugin，江彬教訓 #6 不相容 1.5.x）
# 用 aiohttp 直打 MiniMax t2a_v2 REST API，取得 PCM 後 emit AudioFrame
aiohttp>=3.9.0

# Phase 3：從 Firestore 讀角色 soul + 對話歷史
firebase-admin>=6.5.0

# Phase 3 上 Firestore：firebase-admin
# firebase-admin==6.5.0

# 環境變數
python-dotenv>=1.0.0

# 其他
httpx>=0.26.0
